Transform MPF to RTF

Experimental export to RTF via the r2rtf package

Usage

mpf_to_rtf(
  mpf,
  colwidths = NULL,
  page_type = "letter",
  pg_width = page_dim(page_type)[if (landscape) 2 else 1],
  pg_height = page_dim(page_type)[if (landscape) 1 else 2],
  landscape = FALSE,
  margins = c(4, 4, 4, 4),
  font_size = 8,
  ...
)

Arguments

mpf

MatrixPrintForm. MatrixPrintForm object.

colwidths

character(1). Column widths.

page_type

character(1). Name of a page type. See page_types. Ignored when pg_width and pg_height are set directly.

pg_width

numeric(1). Page width in inches.

pg_height

numeric(1). Page height in inches.

landscape

logical(1). Should the dimensions of page_type be inverted for landscape? Defaults to FALSE, ignored when pg_width and pg_height are set directly.

margins

numeric(4). Named numeric vector containing 'top', 'bottom', 'left', and 'right' margins in inches. Defaults to .5 inches for both vertical margins and .75 for both horizontal margins.

font_size

numeric(1). Font size, defaults to 12.

...

Passed to individual methods.

Value

An rtf object

Details

This function provides a low-level coercion of a MatrixPrintForm object into text containing the corresponding table in RTF`. Currently, no pagination is done at this level, and should be done prior to calling this function, though that may change in the future.