Make row and column layout summary data.frames for use during pagination

Make row and column layout summary data.frames for use during pagination

make_row_df(
  tt,
  colwidths = NULL,
  visible_only = TRUE,
  rownum = 0,
  indent = 0L,
  path = character(),
  incontent = FALSE,
  repr_ext = 0L,
  repr_inds = integer(),
  sibpos = NA_integer_,
  nsibs = NA_integer_
)

Arguments

tt

ANY. Object representing the table-like object to be summarized.

colwidths

numeric. Internal detail do not set manually.

visible_only

logical(1). Should only visible aspects of the table structure be reflected in this summary. Defaults to TRUE. May not be supported by all methods.

rownum

numeric(1). Internal detail do not set manually.

indent

integer(1). Internal detail do not set manually.

path

character. Path to the (sub)table represented by tt. Defaults to character()

incontent

logical(1). Internal detail do not set manually.

repr_ext

integer(1). Internal detail do not set manually.

repr_inds

integer. Internal detail do not set manually.

sibpos

integer(1). Internal detail do not set manually.

nsibs

integer(1). Internal detail do not set manually.

Value

a data.frame of row/column-structure information used by the pagination machinery.

Details

When visible_only is TRUE (the default), methods should return a data.frame with exactly one row per visible row in the table-like object. This is useful when reasoning about how a table will print, but does not reflect the full pathing space of the structure (though the paths which are given will all work as is).

If supported, when visible_only is FALSE, every structural element of the table (in row-space) will be reflected in the returned data.frame, meaning the full pathing-space will be represented but some rows in the layout summary will not represent printed rows in the table as it is displayed.

Most arguments beyond tt and visible_only are present so that `make_row_df` methods can call `make_row_df` recursively and retain information, and should not be set during a top-level call

Note

the technically present root tree node is excluded from the summary returne dby both make_row_df and make_col_df, as it is simply the row/column structure of tt and thus not useful for pathing or pagination.