Create row of pagination data frame

Create row of pagination data frame

Usage

pagdfrow(
  row,
  nm = obj_name(row),
  lab = obj_label(row),
  rnum,
  pth,
  sibpos = NA_integer_,
  nsibs = NA_integer_,
  extent = nlines(row, colwidths),
  colwidths = NULL,
  repext = 0L,
  repind = integer(),
  indent = 0L,
  rclass = class(row),
  nrowrefs = 0L,
  ncellrefs = 0L,
  nreflines = 0L,
  force_page = FALSE,
  page_title = NA_character_,
  trailing_sep = NA_character_
)

Arguments

row

ANY. Object representing the row, which is used for default values of nm, lab, extent and rclass if provided. Must have methods for obj_name, obj_label, and nlines, respectively, for default values of nm, lab and extent to be retrieved, respectively.

nm

character(1). Name

lab

character(1). Label

rnum

numeric(1). Absolute row number

pth

character or NULL. Path within larger table

sibpos

integer(1). Position among sibling rows

nsibs

integer(1). Number of siblings (including self).

extent

numeric(1). Number of lines required to print the row

colwidths

numeric. Column widths

repext

integer(1). Number of lines required to reprint all context for this row if it appears directly after pagination.

repind

integer. Vector of row numbers to be reprinted if this row appears directly after pagination.

indent

integer. Indent

rclass

character(1). Class of row object.

nrowrefs

integer(1). Number of row referential footnotes for this row

ncellrefs

integer(1). Number of cell referential footnotes for the cells in this row

nreflines

integer(1). Total number of lines required by all referential footnotes

force_page

logical(1). Currently Ignored.

page_title

logical(1). Currently Ignored.

trailing_sep

character(1). The string to used as a separator below this row during printing (or NA_character_ for no separator).

Value

a single row data.frame with the columns appropriate for a pagination info data frame.